requirements.txt
django==2.0
djangorestframework==3.8.2
coreapi==2.3.3
python-decouple==3.1